refactor: move version info modal
