+++ frontend/locales/en.json
@@ -156,7 +156,7 @@
     "versionInfo": {
       "issueTracker": "Found a bug? Fill an issue!",
       "sourceCode": "Read the source code",
-      "versionInfo": "Version info",
+      "versionInfo": "Running Version",
       "title": "Running HedgeDoc on version"
     }
   },
